+++ releases.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
   - Add `Async::Limiter::Token#close` to release and permanently invalidate a token.
+  - Expose `Async::Limiter::Generic#timing`.
 
 ## v2.2.0